+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Terminal outcome statements should now correctly exit from iteration
 - MappedHTTPErrors and MappedErrors are handled a little better
 - Parser uses AST metadata to check if source was changed
+- Profile provider resolves integration parameters and fallback to default during `bind`
 
 ### Fixed
 - operation call no longer overwrites already mapped result